+++ extension.js
@@ -35,45 +35,81 @@ function activate(context) {
 
 /**
  * Toggle feature
- * @returns {void}
  */
-function toggleFeature() {
+/**
+ * Toggles a feature in the Rust project.
+ * @returns {Promise<void>} A Promise that resolves when the feature has been toggled.
+ */
+async function toggleFeature() {
   try {
     const config = new ConfigManager();
     const cargoTomlPath = getCargoTomlPath();
     const features = parseCargoToml(cargoTomlPath);
 
-    if (Object.keys(features).length === 0) {
+    if (Object.keys(features).length === 0 && !config.checkFeature('all')) {
       vscode.window.showInformationMessage('No features found');
       return;
     }
 
-    const featureList = Object.keys(features).map((feature) => {
-      return config.checkFeature(feature) ? `[✓] ${feature}` : `[ ] ${feature}`;
+    const featureList = buildFeatureList(config, features);
+    const selectedFeature = await vscode.window.showQuickPick(featureList, {
+      placeHolder: 'Select a feature',
     });
 
-    vscode.window
-      .showQuickPick(featureList, { placeHolder: 'Select a feature' })
-      .then((feature) => {
-        if (feature) {
-          const featureName = feature.slice(4);
-          feature.startsWith('[✓]')
-            ? config.removeFeature(featureName)
-            : config.addFeature(featureName);
-          vscode.window.showInformationMessage(
-            `Feature ${featureName} ${
-              feature.startsWith('[✓]') ? 'disabled' : 'enabled'
-            }`
-          );
-
-          vscode.commands.executeCommand('rust-analyzer.restartServer');
-        }
-      });
+    if (selectedFeature) {
+      await handleSelectedFeature(config, selectedFeature);
+      await vscode.commands.executeCommand('rust-analyzer.restartServer');
+    }
   } catch (error) {
-    console.log(error);
+    console.error(error);
     vscode.window.showErrorMessage(
       'An error occurred while toggling the feature. Please check the console for more details.'
     );
+  }
+}
+
+/**
+ * Builds a list of features based on the given configuration and feature object.
+ *
+ * @param {Object} config - The configuration object.
+ * @param {Object} features - The feature object.
+ * @returns {Array} An array of feature strings with checkboxes indicating whether each feature is enabled or not.
+ */
+function buildFeatureList(config, features) {
+  return [
+    config.checkFeature('all') ? '[✓] all' : '[ ] all',
+    ...Object.keys(features).map((feature) =>
+      config.checkFeature(feature) ? `[✓] ${feature}` : `[ ] ${feature}`
+    ),
+  ];
+}
+
+/**
+ * Handles the selected feature based on the configuration and displays a message to the user.
+ * @param {Object} config - The configuration object.
+ * @param {string} selectedFeature - The selected feature to handle.
+ * @returns {Promise<void>}
+ */
+async function handleSelectedFeature(config, selectedFeature) {
+  const featureName = selectedFeature.slice(4);
+
+  if (featureName === 'all') {
+    if (selectedFeature.startsWith('[✓]')) {
+      config.removeFeature('all');
+      vscode.window.showInformationMessage('All features disabled');
+    } else {
+      config.activateAllFeatures();
+      vscode.window.showInformationMessage('All features enabled');
+    }
+    return;
+  }
+
+  if (selectedFeature.startsWith('[✓]')) {
+    config.removeFeature(featureName);
+    vscode.window.showInformationMessage(`Feature ${featureName} disabled`);
+  } else {
+    config.addFeature(featureName);
+    vscode.window.showInformationMessage(`Feature ${featureName} enabled`);
   }
 }
 
@@ -88,14 +124,19 @@ function updateStatusBarItem(statusBarItem) {
   const features = parseCargoToml(cargoTomlPath);
   const featureListFromSettings = config.getFeatureList();
 
+  if (featureListFromSettings.includes('all')) {
+    statusBarItem.text = '$(gear) All Features Active';
+    statusBarItem.tooltip = 'All features are currently activated.';
+    return;
+  }
+
   if (Object.keys(features).length === 0) {
     statusBarItem.text = 'No features found';
+    statusBarItem.tooltip = 'No features available to activate.';
   } else {
     statusBarItem.text = '$(gear) Toggle Feature';
-    statusBarItem.tooltip = `Enabled Features\n${Object.values(
-      featureListFromSettings
-    )
-      .map((feature) => `[✓] ${feature}`)
+    statusBarItem.tooltip = `Enabled Features\n${featureListFromSettings
+      .map((/** @type {string} */ feature) => `[✓] ${feature}`)
       .join('\n')}`;
   }
 }